myproject/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.6
django-cors-headers==4.8.0
django-phonenumber-field==8.1.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
